+++ edit_mword.php
@@ -172,8 +172,8 @@ else {  // if (! isset($_REQUEST['op']))
     $titletext = ($new ? "New Term" : "Edit Term") . ": " . $term;
     pagestart_nobody($titletext);
     ?>
-<script type="text/javascript" src="js/unloadformcheck.js" charset="utf-8"></script>
 <script type="text/javascript">
+    $(document).ready(ask_before_exiting);
     $(window).on('beforeunload',function() {
         setTimeout(function() {window.parent.frames['ru'].location.href = 'empty.html';}, 0);
     });