+++ new_word.php
@@ -119,11 +119,11 @@ else {  // if (! isset($_REQUEST['op']))
     $scrdir = getScriptDirectionTag($lang);
     pagestart_nobody('');
     ?>
-    <script type="text/javascript" src="js/unloadformcheck.js" charset="utf-8"></script>
     <script type="text/javascript">
-    $(window).on('beforeunload',function() {
-        setTimeout(function() {window.parent.frames['ru'].location.href = 'empty.html';}, 0);
-    });
+        $(document).ready(ask_before_exiting);
+        $(window).on('beforeunload',function() {
+            setTimeout(function() {window.parent.frames['ru'].location.href = 'empty.html';}, 0);
+        });
     </script>
     
     <form name="newword" class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>" method="post">